Removed the exception and all try/catch code that was in the PICS filter. Fixed issues with PICS code on AMR. Also with PICS codes using the correct time slice while processing pathlines. Added code to the PICS code to be able to restart streamlines/pathlines. Also can be used to save the current state to be loaded later. Added code to empty the data queue if using load on demand. Changed the GetMemorySize function to be able to handle memory sizes over 4G.
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@17580 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/common/utility/Utility.h
@@ -59,7 +59,7 @@ int  UTILITY_API  LongestCommonPrefixLength(const char * const *, int);
 int  UTILITY_API  LongestCommonSuffixLength(const char * const *, int);
 
 // LINUX ONLY
-void UTILITY_API  GetMemorySize(unsigned int &, unsigned int &);
+void UTILITY_API  GetMemorySize(unsigned long &, unsigned long &);
 void UTILITY_API  PrintCallStack(ostream &out, const char *file, int line);
 
 void UTILITY_API  WaitUntilFile(const char *);
